fix: Not resend request when hidden document status.
