+++ src/components/ADempiere/Field/index.vue
@@ -33,39 +33,10 @@
         <span v-else key="is-field-name">
           {{ isFieldOnly() }}
         </span>
-        <el-popover
-          v-if="(field.columnName === 'DocStatus') && (!isEmptyValue(processOrdenUuid))"
-          placement="right"
-          width="400"
-          trigger="click"
-        >
-          <el-select
-            v-model="valueActionDocument"
-            @change="documentActionChange"
-          >
-            <el-option
-              v-for="(item, key) in listDocumentActions"
-              :key="key"
-              :label="item.name"
-              :value="item.value"
-            />
-          </el-select>
-          <el-tag
-            v-if="isEmptyValue(valueActionDocument)"
-            :type="tagStatus(field.value)"
-          >
-            {{ field.displayColumn }}
-          </el-tag>
-          <el-tag
-            v-else
-            :type="tagStatus(valueActionDocument)"
-          >
-            {{ labelDocumentActions }}
-          </el-tag>
-          <p v-if="isEmptyValue(valueActionDocument)"> {{ field.description }} </p>
-          <p v-else> {{ descriptionDocumentActions }} </p>
-          <el-button slot="reference" type="text" icon="el-icon-set-up" @click="listActionDocument" />
-        </el-popover>
+        <field-document-status
+          v-if="(field.columnName === 'DocStatus') && (!isEmptyValue(processOrderUuid))"
+          :field="fieldAttributes"
+        />
         <field-translated
           v-if="field.isTranslated && !isAdvancedQuery"
           :field-attributes="fieldAttributes"
@@ -94,6 +65,7 @@
 import FieldContextInfo from '@/components/ADempiere/Field/fieldContextInfo'
 import FieldTranslated from '@/components/ADempiere/Field/fieldTranslated'
 import FieldOperatorComparison from '@/components/ADempiere/Field/fieldOperatorComparison'
+import FieldDocumentStatus from '@/components/ADempiere/Field/fieldPopovers/fieldDocumentStatus'
 import { FIELD_ONLY } from '@/components/ADempiere/Field/references'
 import { DEFAULT_SIZE } from '@/components/ADempiere/Field/fieldSize'
 import { fieldIsDisplayed } from '@/utils/ADempiere'
@@ -108,6 +80,7 @@ export default {
   name: 'FieldDefinition',
   components: {
     FieldContextInfo,
+    FieldDocumentStatus,
     FieldTranslated,
     FieldOperatorComparison
   },
@@ -148,8 +121,7 @@ export default {
   },
   data() {
     return {
-      field: {},
-      valueActionDocument: ''
+      field: {}
     }
   },
   computed: {
@@ -267,36 +239,8 @@ export default {
       }
       return false
     },
-    listDocumentActions() {
-      return this.$store.getters.getListDocumentActions.documentActionsList
-    },
-    defaultDocumentActions() {
-      return this.$store.getters.getListDocumentActions.defaultDocumentAction
-    },
-    labelDocumentActions() {
-      const found = this.listDocumentActions.find(element => {
-        if (element.value === this.valueActionDocument) {
-          return element
-        }
-      })
-      if (this.isEmptyValue(found)) {
-        return this.valueActionDocument
-      }
-      return found.name
-    },
-    descriptionDocumentActions() {
-      const found = this.listDocumentActions.find(element => {
-        if (element.value === this.valueActionDocument) {
-          return element
-        }
-      })
-      if (this.isEmptyValue(found)) {
-        return this.valueActionDocument
-      }
-      return found.description
-    },
-    processOrdenUuid() {
-      return this.$store.getters.getOrden
+    processOrderUuid() {
+      return this.$store.getters.getOrders
     }
   },
   watch: {
@@ -311,41 +255,6 @@ export default {
   },
   methods: {
     showMessage,
-    listActionDocument() {
-      this.$store.dispatch('listDocumentActionStatus', {
-        tableName: 'C_Order',
-        recordUuid: this.$route.query.action
-      })
-    },
-    documentActionChange(value) {
-      var actionProcess = this.$store.getters.getOrden
-      this.$store.dispatch('notifyFieldChange', {
-        parentUuid: this.parentUuid,
-        containerUuid: this.containerUuid,
-        columnName: 'DocAction',
-        isSendToServer: true,
-        newValue: value
-      })
-      this.$store.dispatch('startProcess', {
-        action: {
-          uuid: actionProcess.uuid,
-          id: actionProcess.id,
-          name: actionProcess.name
-        }, // process metadata
-        tableName: this.$route.params.tableName,
-        recordId: this.$route.params.recordId,
-        recordUuid: this.$route.query.action,
-        parametersList: [{
-          columnName: 'DocStatus',
-          value: this.valueActionDocument
-        }],
-        isActionDocument: true,
-        parentUuid: this.parentUuid,
-        panelType: this.panelType,
-        containerUuid: this.containerUuid// determinate if get table name and record id (window) or selection (browser)
-      })
-      this.valueActionDocument = ''
-    },
     isDisplayed() {
       if (this.isAdvancedQuery) {
         return this.field.isShowedFromUser
